Refactor klerosDynamicFormField to work as a Form Field Define an schema and link to klerosDynamicFormField component
